append data connection TLS resumption and make data connection more stable

+++ pftp/logger.go
@@ -9,7 +9,7 @@ import (
 type logger struct {
 	fromip string
 	user   string
-	id     int
+	id     uint64
 }
 
 func (l *logger) debug(format string, args ...interface{}) {
